replace dicts by Enums and add many types to MsgT
The dicts still exists for backwards compability, but are not needed any
more.
The Enum is still a sublass of int for beckwars compability and
for comparisons to byte fields.
The Enums get custom __str__ and __repr__.
A nice Attribute was added to th Enums for a nice human readable
String for some members.